Fix clippy warnings (collapsible_if, unnecessary_cast, type_complexity)

+++ server/src/scanner.rs
@@ -496,8 +496,7 @@ impl EqGameScanner {
         );
 
         let suffix = if match_val != 0 {
-            let current =
-                ir.read_integer_entry("SpawnInfo Offsets", entry.output_label, false) as u64;
+            let current = ir.read_integer_entry("SpawnInfo Offsets", entry.output_label, false);
             if match_val == current {
                 " # Match\r\n"
             } else if write_out {
